fixed project structure + data visualization added
