+++ .gitignore
@@ -216,3 +216,5 @@ __marimo__/
 *.jpg
 project/scripts/data
 project/scripts/output
+project/output
+project/scripts/wandb
