--- project/scripts/config.py
@@ -1,2 +1,0 @@
-# Configuration file for the project
-OUTPUT_DIR = "output"
